Move RawHashFunction to its own module
