+++ src/lib/pk_pad/raw_hash/info.txt
@@ -1,0 +1,12 @@
+<defines>
+RAW_HASH_FN -> 20230221
+</defines>
+
+<module_info>
+name -> "Raw Hash Function"
+brief -> "Hash Function Used For Signature Schemes"
+</module_info>
+
+<requires>
+hash
+</requires>
