+++ src/lib/pubkey/rsa/rsa.cpp
@@ -16,6 +16,7 @@
 #include <botan/internal/monty.h>
 #include <botan/internal/divide.h>
 #include <botan/internal/monty_exp.h>
+#include <botan/internal/emsa.h>
 
 #if defined(BOTAN_HAS_THREAD_UTILS)
   #include <botan/internal/thread_pool.h>
